feat(health): integrate GitHub health data into lesson components
- Added HealthSignalRow component to display GitHub health metrics in LessonPreviewCard and LessonCard.
- Updated LessonRail and PathwayShowcase to pass health data by lesson slug.
- Enhanced CSS for lesson cards to accommodate new health indicators and improve layout.
- Refactored components to utilize the updated health data structure, ensuring better integration and presentation of health information.
